Creating matching algorithm tests
